requirements.txt
numpy
scipy
tqdm
networkx
itertools
torch_geometric
